# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Track Order API",
        "description": "API pour la gestion des commandes.",
        "contact": {
            "email": "ludicagnihouede@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Lister les produits",
                "description": "Récupère une liste paginée de produits avec une option de recherche par nom.",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Terme de recherche pour filtrer les produits par nom.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Laptop"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page (par défaut 10).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des produits.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Laptop Pro"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "High-end laptop"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1200.99
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-26T10:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-26T10:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://example.com/api/products?page=2"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Paramètres invalides."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Ajouter un produit",
                "description": "Permet d'ajouter un nouveau produit avec validation des données.",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom du produit.",
                                        "type": "string",
                                        "example": "Laptop Pro"
                                    },
                                    "description": {
                                        "description": "Description du produit.",
                                        "type": "string",
                                        "example": "Un ordinateur portable haut de gamme."
                                    },
                                    "price": {
                                        "description": "Prix du produit.",
                                        "type": "number",
                                        "format": "float",
                                        "example": 1200.99
                                    },
                                    "stock": {
                                        "description": "Quantité en stock.",
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produit créé avec succès."
                                        },
                                        "product": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Laptop Pro"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Un ordinateur portable haut de gamme."
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1200.99
                                                },
                                                "stock": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "26/03/2025 10:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "26/03/2025 10:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "Le nom est obligatoire."
                                                ],
                                                "price": [
                                                    "Le prix doit être un nombre."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez un token Bearer pour accéder aux routes protégées.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}